moved tests from entry file to a new test file
added test file to package.json test script
