+++ makeall.sh
@@ -1,6 +1,6 @@
 (
 echo 'function lamejs() {'
-browserify --list src/js/lametest.js  | grep lamejs | grep -v node_modules | xargs cat | grep -v -e 'common\..*;' -e 'require(' -e 'module.exports.*;$' | sed 's/^module.exports = {/var module_exports = {/';
+browserify --list src/js/index.js  | grep lamejs | grep -v node_modules | xargs cat | grep -v -e 'common\..*;' -e 'require(' -e 'module.exports.*;$' | sed 's/^module.exports = {/var module_exports = {/';
 echo 'L3Side.SFBMAX = (Encoder.SBMAX_s * 3);'
 echo '//testFullLength();'
 echo 'lamejs.Mp3Encoder = Mp3Encoder;'
